+++ superstore/src/commonTest/kotlin/org/mobilenativefoundation/store/superstore5/util/TestWarehouse.kt
@@ -1,0 +1,7 @@
+package org.mobilenativefoundation.store.superstore5.util
+
+import org.mobilenativefoundation.store.superstore5.Warehouse
+
+interface TestWarehouse<Key : Any, Output : Any> : Warehouse<Key, Output> {
+    suspend fun get(key: String, fail: Boolean = false, ttl: Long? = null): Output?
+}
